backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
openai>=1.52.0
pinecone==4.1.0
langchain==0.2.6
langchain-community==0.2.6
sentence-transformers==3.0.1
# PyMuPDF==1.24.5
PyMuPDF==1.23.26
numpy==1.26.4
gunicorn==22.0.0
